+++ csv_importer/alembic.ini
@@ -63,7 +63,9 @@ version_path_separator = os  # Use os.pathsep. Default configuration used for ne
 
 # NB for relative paths use three slashes -> ///
 #    for absolute paths use four slashes -> ////
-sqlalchemy.url = sqlite:///sqlite/data.db
+# sqlalchemy.url = sqlite:///sqlite/data.db
+
+sqlalchemy.url = postgresql://postgres:data-for-good@localhost/eddata
 
 
 [post_write_hooks]
