--- csv_importer/alembic/versions/2024_04_28_1155-first_version.py
@@ -1,55 +1,0 @@
-"""first version
-
-Revision ID: beec913ff706
-Revises: 
-Create Date: 2024-04-28 11:55:29.192977
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'beec913ff706'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('region',
-    sa.Column('id', sa.Integer(), sa.Identity(always=False), nullable=False),
-    sa.Column('name', sa.String(length=20), nullable=True)
-    )
-    op.create_table('municipality',
-    sa.Column('id', sa.Integer(), sa.Identity(always=False), nullable=False),
-    sa.Column('name', sa.String(length=20), nullable=True),
-    sa.Column('region_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['region_id'], ['region.id'], )
-    )
-    op.create_table('place',
-    sa.Column('id', sa.Integer(), sa.Identity(always=False), nullable=False),
-    sa.Column('name', sa.String(length=20), nullable=True),
-    sa.Column('municipality_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['municipality_id'], ['municipality.id'], )
-    )
-    op.create_table('school',
-    sa.Column('id', sa.String(length=10), nullable=False),
-    sa.Column('name', sa.String(length=30), nullable=False),
-    sa.Column('place_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['place_id'], ['place.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('school')
-    op.drop_table('place')
-    op.drop_table('municipality')
-    op.drop_table('region')
-    # ### end Alembic commands ###
