--- csv_importer/alembic/versions/2024_04_28_1424-add_examination_tables.py
@@ -1,73 +1,0 @@
-"""add examination tables
-
-Revision ID: 230c17fd2335
-Revises: beec913ff706
-Create Date: 2024-04-28 14:24:55.243648
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '230c17fd2335'
-down_revision: Union[str, None] = 'beec913ff706'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('examination',
-    sa.Column('id', sa.String(length=15), nullable=True),
-    sa.Column('type', sa.String(length=5), nullable=True),
-    sa.Column('year', sa.Integer(), nullable=True),
-    sa.Column('grade_level', sa.Integer(), nullable=True)
-    )
-    op.create_table('examination_score',
-    sa.Column('examination_id', sa.String(length=15), nullable=False),
-    sa.Column('school_id', sa.String(length=10), nullable=False),
-    sa.Column('subject', sa.String(length=10), nullable=False),
-    sa.Column('subject_specifier', sa.String(length=10), nullable=False),
-    sa.Column('people', sa.Integer(), nullable=True),
-    sa.Column('score', sa.Double(), nullable=True),
-    sa.Column('max_possible_score', sa.Double(), nullable=True),
-    sa.ForeignKeyConstraint(['examination_id'], ['examination.id'], ),
-    sa.ForeignKeyConstraint(['school_id'], ['school.id'], ),
-    sa.PrimaryKeyConstraint('examination_id', 'school_id', 'subject', 'subject_specifier')
-    )
-    # TODO: why alembic generates this?!?!
-    # op.alter_column('municipality', 'id',
-    #            existing_type=sa.INTEGER(),
-    #            server_default=sa.Identity(always=False),
-    #            existing_nullable=False)
-    # op.alter_column('place', 'id',
-    #            existing_type=sa.INTEGER(),
-    #            server_default=sa.Identity(always=False),
-    #            existing_nullable=False)
-    # op.alter_column('region', 'id',
-    #            existing_type=sa.INTEGER(),
-    #            server_default=sa.Identity(always=False),
-    #            existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('region', 'id',
-               existing_type=sa.INTEGER(),
-               server_default=None,
-               existing_nullable=False)
-    op.alter_column('place', 'id',
-               existing_type=sa.INTEGER(),
-               server_default=None,
-               existing_nullable=False)
-    op.alter_column('municipality', 'id',
-               existing_type=sa.INTEGER(),
-               server_default=None,
-               existing_nullable=False)
-    op.drop_table('examination_score')
-    op.drop_table('examination')
-    # ### end Alembic commands ###
