+++ csv_importer/alembic/versions/2024_05_01_1041-create_tables.py
@@ -1,0 +1,79 @@
+"""Create tables
+
+Revision ID: 200d6e6d7372
+Revises: 
+Create Date: 2024-05-01 10:41:17.105127
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '200d6e6d7372'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('examination',
+    sa.Column('id', sa.String(length=15), nullable=False),
+    sa.Column('type', sa.String(length=5), nullable=True),
+    sa.Column('year', sa.Integer(), nullable=True),
+    sa.Column('grade_level', sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('region',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(length=20), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('municipality',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(length=20), nullable=True),
+    sa.Column('region_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['region_id'], ['region.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('place',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.String(length=40), nullable=True),
+    sa.Column('municipality_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['municipality_id'], ['municipality.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('school',
+    sa.Column('id', sa.String(length=10), nullable=False),
+    sa.Column('name', sa.String(length=150), nullable=False),
+    sa.Column('place_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['place_id'], ['place.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('examination_score',
+    sa.Column('examination_id', sa.String(length=15), nullable=False),
+    sa.Column('school_id', sa.String(length=10), nullable=False),
+    sa.Column('subject', sa.String(length=10), nullable=False),
+    sa.Column('subject_specifier', sa.String(length=10), nullable=False),
+    sa.Column('people', sa.Integer(), nullable=True),
+    sa.Column('score', sa.Double(), nullable=True),
+    sa.Column('max_possible_score', sa.Double(), nullable=True),
+    sa.ForeignKeyConstraint(['examination_id'], ['examination.id'], ),
+    sa.ForeignKeyConstraint(['school_id'], ['school.id'], ),
+    sa.PrimaryKeyConstraint('examination_id', 'school_id', 'subject', 'subject_specifier')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('examination_score')
+    op.drop_table('school')
+    op.drop_table('place')
+    op.drop_table('municipality')
+    op.drop_table('region')
+    op.drop_table('examination')
+    # ### end Alembic commands ###
