+++ csv_importer/models.py
@@ -8,14 +8,14 @@ Models = MetaData()
 Region: Table = Table(
     'region',
     Models,
-    Column('id', Integer, Identity()),
+    Column('id', Integer, primary_key=True),
     Column('name', String(20))
 )
 
 Municipality: Table = Table(
     'municipality',
     Models,
-    Column('id', Integer, Identity()),
+    Column('id', Integer, primary_key=True),
     Column('name', String(20)),
     Column('region_id', ForeignKey('region.id'), nullable=False)
 )
@@ -23,8 +23,8 @@ Municipality: Table = Table(
 Place: Table = Table(
     'place',
     Models,
-    Column('id', Integer, Identity()),
-    Column('name', String(20)),
+    Column('id', Integer, primary_key=True),
+    Column('name', String(40)),
     Column('municipality_id', ForeignKey('municipality.id'), nullable=False)
 )
 
@@ -32,7 +32,7 @@ School: Table = Table(
     'school',
     Models,
     Column('id', String(10), primary_key=True),
-    Column('name', String(30), nullable=False),
+    Column('name', String(150), nullable=False),
     Column('place_id', ForeignKey('place.id'), nullable=False)
 )
 
@@ -40,7 +40,7 @@ Examination: Table = Table(
     'examination',
     Models,
     # id will be in the form: nvo-4-2022
-    Column('id', String(15)),
+    Column('id', String(15), primary_key=True),
 
     # type will be НВО or ДЗИ
     Column('type', String(5)),
