+++ csv_importer/postgres/.gitignore
@@ -1,0 +1,1 @@
+pgdata/
